+++ CHANGELOG.md
@@ -9,8 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added `OpenXmlElementFunctionalExtensions.With` extension methods, which offer flexible means for constructing `OpenXmlElement` instances in the context of pure functional transformations.
 
 ### Fixed
-- Ensures attributes are available when `OpenXmlElement` is initialized with outer XML (#684)
+- Ensures attributes are available when `OpenXmlElement` is initialized with outer XML (#684, #692)
 - Some documentation errors (#681)
+- Removed state that made it non-threadsafe to validate elements under certain conditions (#686)
+- Correctly inserts strongly-typed elements before known elements that are not strongly-typed (#690)
 
 ## Version 2.10.0 - 2020-01-10
 ### Added