+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Version 2.10.1 - 2020-02-26
+## Version 2.10.1 - 2020-02-28
 
 ### Fixed
 - Ensures attributes are available when `OpenXmlElement` is initialized with outer XML (#684, #692)